split into `context/` subdir
